feat(web-core): standardize console component between xo-lite and xo-6
